+++ test/td_cache/concept_cache_test.exs
@@ -32,7 +32,6 @@ defmodule TdCache.ConceptCacheTest do
         "business_concept:ids:active",
         "business_concept:ids:inactive",
         "business_concept:ids:confidential",
-        "domain:events",
         "domain:deleted_ids"
       ])
     end)