+++ test/td_cache/domain_cache_test.exs
@@ -6,7 +6,6 @@ defmodule TdCache.DomainCacheTest do
 
   alias TdCache.DomainCache
   alias TdCache.Redix
-  alias TdCache.Redix.Stream
 
   doctest TdCache.DomainCache
 
@@ -30,17 +29,6 @@ defmodule TdCache.DomainCacheTest do
       assert_structs_equal(d, domain, [:id, :name, :external_id])
     end
 
-    test "publishes an event when a domain is created", %{domain: domain} do
-      {:ok, _} = DomainCache.put(domain)
-      assert {:ok, events} = Stream.read(:redix, ["domain:events"], transform: true)
-      assert [%{event: "domain_created"}] = events
-    end
-
-    test "does not publish an event if publish: false is specified", %{domain: domain} do
-      {:ok, _} = DomainCache.put(domain, publish: false)
-      assert {:ok, []} = Stream.read(:redix, ["domain:events"], transform: true)
-    end
-
     test "updates a domain entry only if changed", %{domain: domain} do
       assert {:ok, [3, 1, 1, 1, 1, 0]} = DomainCache.put(domain)
       assert {:ok, []} = DomainCache.put(domain)
@@ -49,9 +37,6 @@ defmodule TdCache.DomainCacheTest do
 
       assert {:ok, [0, 0, 0, 0, 0, 0]} = DomainCache.put(updated)
       assert {:ok, %{name: "updated name"}} = DomainCache.get(domain.id)
-
-      assert {:ok, events} = Stream.read(:redix, ["domain:events"], transform: true)
-      assert [%{event: "domain_created"}, %{event: "domain_updated"}] = events
     end
 
     test "deletes an entry in redis", %{domain: %{id: id} = domain} do